+++ src/TwigComponent/CHANGELOG.md
@@ -8,8 +8,8 @@
 
 -   Add `PostMount` hook component hook to intercept extra props.
 
--   Add `HasAttributesTrait` for components which makes `attributes` variable available
-    in component templates.
+-   Add attributes system that takes extra props passed to `component()` and converts them
+    into a `ComponentAttributes` object available in your template as `attributes`.
 
 -   Add `PreRenderEvent` to intercept/manipulate twig template/variables before rendering.